+++ libs/docusaurus-plugin-docgen/jest.config.mjs
@@ -1,0 +1,6 @@
+/** @type {import('@jest/types').Config.InitialOptions} */
+export default {
+  preset: '../../jest.preset.js',
+  displayName: 'docusaurus-plugin-docgen',
+  testEnvironment: 'node',
+};